Modal Factory - Remove second call to componentDidMount
There were two instances of the componentDidMount function inside of the the modal factory.
This caused the first instance of the function to be overwritten.
The effect of this is that the onShow callback would never get called.

Moved the keydown listener functionality into the first declaration of the function, and removed
the second declaration.